fix(stress): satisfy gosec/unconvert lints in stressgen helpers
CI golangci-lint flagged three issues in the freshly-added stressgen
package:

- G115 + G602 in writeRandomFile: the manual 8-byte unpack tripped both
  the integer-overflow and slice-out-of-range checkers. Switch to
  encoding/binary.LittleEndian.PutUint64 over an `i+8 <= bufSize` loop
  bound — same output, no manual indexing, gosec sees the safety.
- G115 in writeRandomJPEG: explicit `& 0xff` masks on the per-channel
  byte conversions so the truncation is intentional and visible.
- unconvert in EnsureFreeSpace: `uint64(st.Bavail)` was a no-op on
  Linux (Bavail is already uint64); drop the cast, keep the //nolint
  on the Bsize multiplication where it's still required.

`go test -race ./internal/tools/...` and `-bench=BenchmarkRun
-benchtime=1x ./internal/tools/hash/` both stay green.

+++ internal/tools/testutil/stressgen/stressgen.go
@@ -33,7 +33,7 @@ func EnsureFreeSpace(dir string) error {
 	if err := syscall.Statfs(dir, &st); err != nil {
 		return nil
 	}
-	free := uint64(st.Bavail) * uint64(st.Bsize) //nolint:gosec // values fit
+	free := st.Bavail * uint64(st.Bsize) //nolint:gosec // values fit
 	if free < MinFreeBytes {
 		return fmt.Errorf("stressgen: need at least %d GiB free at %s, have %d MiB",
 			MinFreeBytes>>30, dir, free>>20)
